Feature/set password

+++ backend/fan_zone/props/used/offers/models.py
@@ -15,4 +15,4 @@ class Offer(models.Model):
     on_delete=models.CASCADE,
     related_name='+'
   )
-  amount = models.PositiveIntegerField()
+  amount = models.FloatField()
